+++ src/platform/EFR32/OTAImageProcessorImpl.h
@@ -55,7 +55,7 @@ private:
      */
     CHIP_ERROR ReleaseBlock();
 
-    static uint16_t mWriteOffset; // End of last written block
+    static uint32_t mWriteOffset; // End of last written block
     static uint8_t mSlotId;       // Bootloader storage slot
     MutableByteSpan mBlock;
     OTADownloader * mDownloader;
